+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.4.0 (2022-02-24)
+
+- Add P2WPKH-P2SH addresses ([#3](https://github.com/ayrat555/cryptopunk/pull/3))
+- Add bech32 addresses ([#4](https://github.com/ayrat555/cryptopunk/pull/4))
+
 ## 0.3.1 (2022-02-12)
 
 - Update ex_pbkdf2 and ex_secp256k1 ([#2](https://github.com/ayrat555/cryptopunk/pull/2))